// apps/core/src/common/enums/page.enum.ts
export enum Page {
  INDEX = 'index',
  HOTEL = 'hotel',
  WELCOME_TO_HABBO_HOTEL = 'welcome_to_habbo_hotel',
  HOW_TO_GET_STARTED = 'how_to_get_started',
  CHATTING = 'chatting',
  HELP_SAFETY = 'help_safety',
  MEETING = 'meeting',
  NAVIGATOR = 'navigator',
  ROOM = 'room',
  FURNITURE = 'furniture',
  CATALOGUE = 'catalogue',
  CATALOGUE1 = 'catalogue1',
  CATALOGUE2 = 'catalogue2',
  CATALOGUE4 = 'catalogue4',
  DECORATION_EXAMPLES = 'decoration_examples',
  STAFF_FAVOURITE = 'staff_favourite',
  TRADING = 'trading',
  PETS = 'pets',
  TAKING_CARE_OF_YOUR_PET = 'taking_care_of_your_pet',
  NEWS = 'news',
  HOME = 'home',
  HOMEPAGE = 'homepage',
  CLIENT = 'client',
  LOGIN = 'login',
  LOGOUT = 'logout',
  PREFERENCES = 'preferences',
  LOOK = 'look',
  CREDITS = 'credits',
  PAYPAL = 'paypal',
  USE_MY_CREDITS = 'use_my_credits',
  CLUB = 'club',
  HABBO_CLUB_BENEFITS = 'habbo_club_benefits',
  JOIN_HABBO_CLUB = 'join_habbo_club',
  HABBO_CLUB_FURNI = 'habbo_club_furni',
  IOT = 'iot',
  REGISTER = 'register',
}